Monodevelop: add mono-devel runtime dep
Summary:
The other day I imported a project in monodevelop, but the IDE was not able to download
the required deps because I was missing the certificates. Those certs are inside mono-devel.
Also, being an IDE, it actually makes sense to require the -devel package of mono.

Test Plan: Said dependencies were downloaded once I had installed mono-devel.

Reviewers: #triage_team, sunnyflunk

Reviewed By: #triage_team, sunnyflunk

Subscribers: sunnyflunk

Differential Revision: https://dev.getsol.us/D4219